feat: refactor buyer profile into smaller components
